make hailo-drivers fit into the RHEL 9 backport
RHEL kernels are designed for long-term API stability and therefore keep
the kernel version unchanged throughout the lifetime of a major RHEL
release. This means that numerous backports of fixes and enhancements
from newer upstream kernels must be integrated into these RHEL kernels.

This patch fixes issues with mismatched precompiler directives to match
the RHEL 9 kernel situation. It is not meant to be applied to the
hailo-drivers source in the upstream git repo, but it should be used in
conjunction with the appropriate RPM spec file to create an installable
RPM package.

Signed-off-by: Sebastian Hetze <she@redhat.com>